+++ src/main/java/profect/group1/goormdotcom/product/infrastructure/client/StockService/StockClient.java
@@ -15,6 +15,7 @@ import java.util.UUID;
 
 @FeignClient(
     name = "product-to-stock",
+    url = "${service.order.url}",
     fallback = StockClientFallback.class,
     configuration = FeignConfig.class
 )
