+++ src/main/java/profect/group1/goormdotcom/review/infrastructure/client/PresignedClient.java
@@ -1,17 +1,17 @@
-package profect.group1.goormdotcom.review.service;
+package profect.group1.goormdotcom.review.infrastructure.client;
 
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
+import profect.group1.goormdotcom.review.infrastructure.client.dto.ObjectKeyResponseDto;
 
 import java.util.UUID;
 @FeignClient(
-        name = "review-to-presigned"
+        name = "review-to-presigned",
+        url = "${service.order.url}"
 )
 public interface PresignedClient {
